requirements.txt
# Machine Learning & Data Processing
scikit-learn>=1.3.0
pandas>=2.0.3
numpy>=1.24.3
joblib>=1.3.2

# Network Analysis
scapy>=2.5.0
pyshark>=0.6

# Web Framework
Flask>=2.3.2
Flask-SQLAlchemy>=3.0.5
Flask-Login>=0.6.2
Flask-WTF>=1.1.1

# Database
psycopg2-binary>=2.9.6
SQLAlchemy>=2.0.19

# Visualization
plotly>=5.15.0

# Real-time Updates
Flask-SocketIO>=5.3.5
eventlet>=0.33.3

# System & Utils
python-dotenv>=1.0.0
schedule>=1.2.0

